chore: use Nexus public group
This configures `settings.xml` exclusively use Nexus Public group that
groups repositories that we use (and a bunch more) into a single
repository.

It reconfigures Nexus to use `nexus.xml` from ConfigMap via `subpath`
volumeMount to `/usr/local/share/nexus.xml` which in turn is copied over
on first start by `fabric8/nexus` Docker image. As the current
`nexus.xml` is already copied over into `/sonatype-work/conf/nexus.xml`
it will not be applied to a already used Nexus persisted volume it
either needs to be removed (the file), or the volume re-created.